Configure Emacs to work with Chemacs2.
This commit changes where my handrolled `.emacs.d` configuration is
saved.  It is now symlinked to `$HOME/.handrolled-emacs` so that
Chemacs2 can be cloned into `$HOME/.emacs.d`.  This allows me to also
make use of Doom Emacs on the same system and swap between the two
configurations with ease. Doom Emacs is expected to be cloned to
`$HOME/.doom-emacs`.