Add BP dev instance tags to PS1 variable.
Exports a new environment variable `TAGFILE`, which points to a text
file that contains the tag of the AWS instance set during creation.
The tag is only set once if the file is not present, and if the
current host does not match the pattern of a BP dev server then the
file is not created.

Currently the previous command indication does not work. I switched
away from using the `prev_func_err` function because I assumed that
the call to `cat $TAGFILE` when setting the `PS1` variable was causing
the `$?` variablet to always be 0. What I have currently still does
not work.