Use `exec-path-from-shell` to gain access to Mac OS path variable.
This allows for ripgrep to be used for regexp searches within
`project`, and also allows for access to the system binary of
`aspell`, which we can now use as our ispell program.

Setting the ispell program has to occur after `exec-path-from-shell`
is configured as to have access to the system path variable (and thus
the location of the `aspell` binary).  There might be a way to
configure this dependency via `use-package`, but currently setting the
ispell program is done in-line within the init.el file, so I just
moved it to the end of the file to ensure that the `use-package` setup
for `exec-path-from-shell` occurs before it is reached.