Conditionally attempt to source bash configs.
This is a quick fix for an issue with how dotfiles are installed on
GitHub codespaces which results in directories not being copied
correctly to the .config home directory.  If the bash directory is not
present, then it throws errors anytime the .bashrc is sourced.

Also since the configuration directory for bash is now dynamic, it
needs to be ignored for shellcheck purposes.  Review commit
8c9544596e0c13137773cc3610f0a1adda3cb06e for additional
details.